% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_duplicates.R
\name{clean_duplicates}
\alias{clean_duplicates}
\title{Clean Duplicate Time Entries}
\usage{
clean_duplicates(con)
}
\arguments{
\item{con}{A database connection object.}
}
\value{
An integer number of duplicate rows found.
}
\description{
Clean Duplicate Time Entries
}
