% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_payroll_start_end_dates.R
\name{get_payroll_start_end_dates}
\alias{get_payroll_start_end_dates}
\title{Payroll Start and End Dates}
\usage{
get_payroll_start_end_dates(todays_date = Sys.Date())
}
\arguments{
\item{todays_date}{}
}
\value{
A list of date strings.
}
\description{
Determine the beginning and end of a payroll period.
}
