% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timesheet_hours.R
\name{get_timesheet_hours}
\alias{get_timesheet_hours}
\title{Get Timesheet Hours}
\usage{
get_timesheet_hours(timesheet_path)
}
\arguments{
\item{timesheet_path}{A string. The path to the timesheet. See
Admin > Personnel > Timesheets > Timesheet Template for an example.}
}
\value{
A data frame with employee, job number, hours worked, and dates
worked.
}
\description{
This function converts a Balcones timesheet into a tidy data
frame.
}
