% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_time.R
\name{upload_time}
\alias{upload_time}
\title{Upload Time}
\usage{
upload_time(timesheet, dev_mode = TRUE)
}
\arguments{
\item{timesheet}{A tidy data frame of employee time. The result of
get_timesheet_hours()}

\item{dev_mode}{A logical. Should the upload be to a temporary table for
development purposes only.}
}
\value{
Nothing is returned.
}
\description{
This function connects to the MySQL instance and uploads the
timesheet data.
}
